requirements.txt
Flask>=2.3,<4
python-dotenv>=1.0,<2
gunicorn>=21,<23
